### 2020-01-25 #2040 @whedon
[PRE REVIEW]: SpiceyPy: a Pythonic Wrapper for the SPICE Toolkit
**Submitting author:** @andrewannex (<a href="http://orcid.org/0000-0002-0253-2313">Andrew Annex</a>)
**Repository:** <a href="https://github.com/AndrewAnnex/SpiceyPy" target ="_blank">https://github.com/AndrewAnnex/SpiceyPy</a>
**Version:** v3.0.1
