+++ joss/paper.bib
@@ -37,7 +37,7 @@
   author={Stansby, David and Rai, Yatharth and Broll, Jeffrey and Shaw, Siddhant and others},
   journal={Astrophysics Source Code Library},
   doi={10.5281/zenodo.1009079},
-  url={https://ascl.net/1903.017},
+  url={https://zenodo.org/record/3540696},
   year={2019}
 }
 
