+++ joss/paper.bib
@@ -70,7 +70,7 @@
   number={3},
   pages={919--930},
   year={2018},
-  doi = {10.2514/1.a34329},
+  doi={10.2514/1.a34329},
   url={https://doi.org/10.2514/1.a34329},
   publisher={American Institute of Aeronautics and Astronautics}
 }
