### 2020-01-26 1acb4c1
added dois (#341)
* Update paper.bib to have more dois
