+++ joss/paper.bib
@@ -36,6 +36,8 @@
   title={HelioPy: Heliospheric and planetary physics library},
   author={Stansby, David and Rai, Yatharth and Broll, Jeffrey and Shaw, Siddhant and others},
   journal={Astrophysics Source Code Library},
+  doi={10.5281/zenodo.1009079},
+  url={https://zenodo.org/record/3540696},
   year={2019}
 }
 
@@ -43,6 +45,7 @@
   title={Applying artificial intelligence to space communications networks: Cognitive real-time link layer adaptations through rapid orbit planning},
   author={Hackett, Timothy Michael},
   year={2019},
+  url={https://etda.libraries.psu.edu/catalog/16634tmh5344},
   school={The Pennsylvania State University}
 }
 
@@ -54,6 +57,8 @@
   number={12},
   pages={10--384},
   year={2018},
+  doi={10.1029/2018ja025877},
+  url={https://doi.org/10.1029/2018ja025877},
   publisher={Wiley Online Library}
 }
 
@@ -65,6 +70,8 @@
   number={3},
   pages={919--930},
   year={2018},
+  doi={10.2514/1.a34329},
+  url={https://doi.org/10.2514/1.a34329},
   publisher={American Institute of Aeronautics and Astronautics}
 }
 
@@ -76,6 +83,8 @@
   number={4},
   pages={1411--1418},
   year={2016},
+  doi={10.1002/2015gl067436},
+  url={https://doi.org/10.1002/2015gl067436},
   publisher={Wiley Online Library}
 }
 
@@ -83,6 +92,8 @@
   title={Machine learning and Monte Carlo based data analysis methods in cosmic dust research},
   author={Albin, Thomas},
   school={University of Stuttgart},
+  doi={10.18419/opus-10454},
+  url={https://doi.org/10.18419/opus-10454},
   year={2019}
 }
 
@@ -91,6 +102,8 @@
   author={Attree, Nicholas and Jorda, Laurent and Groussin, Olivier and Mottola, Stefano and Thomas, Nicolas and Brouet, Yann and K{\"u}hrt, Ekkehard and Knapmeyer, Martin and Preusker, Frank and Scholten, Frank and others},
   journal={Astronomy and astrophysics},
   year={2019},
+  doi={10.1051/0004-6361/201834415},
+  url={https://doi.org/10.1051/0004-6361/201834415},
   publisher={EDP Sciences}
 }
 
@@ -102,6 +115,8 @@
   number={Suppl\_2},
   pages={S396--S403},
   year={2017},
+  doi={10.1093/mnras/stx1871},
+  url={https://doi.org/10.1093/mnras/stx1871},
   publisher={Oxford University Press}
 }
 
@@ -112,6 +127,8 @@
   volume={10000},
   pages={1000011},
   year={2016},
+  doi={10.1117/12.2240648},
+  url={https://doi.org/10.1117/12.2240648},
   organization={International Society for Optics and Photonics}
 }
 
@@ -120,6 +137,10 @@
   author={Costa, Marc and Grass, Markus},
   booktitle={European Planetary Science Congress},
   volume={12},
+  month={Sep},
+  eid={EPSC2018-53},
+  pages={EPSC2018-53},
+  url={https://ui.adsabs.harvard.edu/abs/2018EPSC...12...53C},
   year={2018}
 }
 
