### 2020-01-26 #2040 @AndrewAnnex
updated paper.bib to include those missing dois...
