### 2020-01-26 #2040 @AndrewAnnex
hmm, well I updated the master branch paper.bib file to include the above dois
