### 2020-01-29 c2ed0c0
datetime2et: handle timezones correctly (#342)
* datetime2et: handle timezone-aware datetimes correctly 
* CSPICE always expects a UTC format without timezone information
* now timezone-aware datetimes are converted to UTC if needed and then the tzinfo is dropped before passing on to SPICE in datetime2et
