### 2020-01-29 78bb9d4
Update README.rst
added black to badges
