### 2020-02-01 4966b6a
fix urllib error in gettestkernels that can happen, some black reformating (#345)
