### 2020-02-01 55d2dbf
Fix internal reference in docs (#344)
Labels that are not placed before a section title need an explicit
title
