### 2020-02-01 d4a9b8e
fixing some sphinx warnings (#347)
