+++ docs/insitu_sensing.rst
@@ -1,5 +1,5 @@
-In-situ Sensing Hands-On Lesson, using CASSINI (Python)
-=======================================================
+In-situ Sensing Hands-On Lesson, using CASSINI
+===============================================
 
 November 20, 2017
 
@@ -186,29 +186,29 @@ ephemeris seconds past J2000, .
 --------------------
 
 Find out what SPICE kernel(s) is(are) needed to support this conversion.
-Reference the "time.req" and/or" Time" tutorial.
+Reference the "time.req" and/or "Time" tutorial.
 
 Find necessary kernel(s) on the NAIF's FTP site.
 
 Find out what routine should be called to load necessary kernel(s).
-Reference the "kernel.req" and/or" Loading Kernels" tutorial.
+Reference the "kernel.req" and/or "Loading Kernels" tutorial.
 
 Find the
-"loader" routine calling sequence specification. Look at the" time.req"and
+"loader" routine calling sequence specification. Look at the "time.req"and
 that routine's source code header. This routine may be an entry point,
 in which case there will be no source file with the same name. To find
 out in which source file this entry point is, search for its name in the
 "Permuted Index".
 
 Find the routine(s) used to convert time between UTC and ET. Look at the
-"time.req" and/or" Time" tutorial.
+"time.req" and/or "Time" tutorial.
 
 Find the
-"converter" routine(s) calling sequence specification. Look    in the" time.req"
+"converter" routine(s) calling sequence specification. Look in the "time.req"
 and the routine's source code header.
 
 Put all calls together in a program, add variable declarations (the
-routine header's "Declarations" and" Examples" sections are a good
+routine header's "Declarations" and "Examples" sections are a good
 place to look for declaration specification and examples) and output
 print statements.
 
@@ -250,7 +250,7 @@ or it can be listed in a meta-kernel:
       \begintext
 
 the name of which, let's call it "convrt.tm", can be then provided as
-a sole argument of the :py:meth:`spiceypy.spiceypy.furnsh` routine:
+a sole argument of the :py:func:`spiceypy.spiceypy.furnsh` routine:
 
 .. code-block:: python
 
@@ -264,10 +264,10 @@ simply adding more kernels to the list in KERNEL_TO_LOAD without
 changing the program code will accomplish that.
 
 The highest level SpiceyPy time routine converting UTC to ET is
-spiceypy.str2et :py:meth:`spiceypy.spiceypy.str2et` .
+spiceypy.str2et :py:func:`spiceypy.spiceypy.str2et` .
 
 It has two arguments – input time string representing UTC in a variety
-of formats (see spiceypy.str2et header's section "Particulars" for
+of formats (see :py:func:`spiceypy.spiceypy.str2et` header's section "Particulars" for
 the complete description of input time formats) and output DP number of
 ET seconds past J2000. A call to spiceypy.str2et converting a given UTC
 to ET could look like this:
@@ -277,7 +277,7 @@ to ET could look like this:
           utc =  '2004-06-11T19:32:00'
           et = spiceypy.str2et(utc)
 
-By combining spiceypy.furnsh and spiceypy.str2et calls and required
+By combining :py:func:`spiceypy.spiceypy.furnsh` and :py:func:`spiceypy.spiceypy.str2et` calls and required
 declarations and by adding a simple print statement, one would get a
 complete program that prints ET for the given UTC epoch.
 
@@ -355,17 +355,17 @@ CASSINI on-board clock epoch "1465674964.105".
 
 Find out what additional (to those already loaded in Step-1) SPICE
 kernel(s) is(are) needed to support SCLK to ET conversion. Look at the
[…]
