### 2020-02-01 ab54094
fix author name spelling
fixes #346
