### 2020-02-15 #2050 @jessie-dotson
My review is complete.  Recommend accepting as is.

I identified a couple of modest changes that would make the documentation more beginner friendly and filed those as issues in the SpiceyPy repository (#350 and #351).  These are not blockers in anyway, just suggestions to the author.

(An aside to JOSS editors...  I did confirm some functionality, but the package includes so many useful functions, it wasn't feasible for me to confirm them all.  As to whether or not the functionality I did confirm was "core functionality" would depend on why you like spice...  )
