### 2020-02-15 b5a004a
doc heading fixes and adding links to required reading, user guides, and permuted index
