+++ docs/remote_sensing.rst
@@ -30,6 +30,7 @@ provided in binary kernels—SPKs and CKs—can be viewed using
 Toolkit installation tree.
 
 Tutorials
+^^^^^^^^^^
 
 The following SPICE tutorials serve as references for the discussions in
 this lesson:
@@ -54,6 +55,11 @@ These tutorials are available from the NAIF ftp server at JPL:
       https://naif.jpl.nasa.gov/naif/tutorials.html
 
 Required Readings
+^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Required Readings <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html>`_ are also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html.
 
 The Required Reading documents are provided with the Toolkit and are
 located under the "cspice/doc" directory in the CSPICE Toolkit
@@ -73,6 +79,12 @@ installation tree.
       time.req         Time conversion
 
 The Permuted Index
+^^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Permuted Index <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html>`_ is also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html.
+
 
 Another useful document distributed with the Toolkit is the permuted
 index. This is located under the "cspice/doc" directory in the C
@@ -83,6 +95,7 @@ discover which SpiceyPy functions perform functions of interest, as well
 as the names of the source files that contain these functions.
 
 SpiceyPy API Documentation
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 A SpiceyPy function's parameters specification is available using the
 built-in Python help system.
@@ -217,7 +230,7 @@ Time Conversion (convtm)
 ------------------------------
 
 Task Statement
---------------
+^^^^^^^^^^^^^^
 
 Write a program that prompts the user for an input UTC time string,
 converts it to the following time systems and output formats:
@@ -234,14 +247,14 @@ and displays the results. Use the program to convert “2004 jun 11
 19:32:00” UTC into these alternate systems.
 
 Learning Goals
---------------
+^^^^^^^^^^^^^^
 
 Familiarity with the various time conversion and parsing functions
 available in the Toolkit. Exposure to source code headers and their
 usage in learning to call functions.
 
 Approach
---------
+^^^^^^^^
 
 The solution to the problem can be broken down into a series of simple
 steps:
@@ -271,7 +284,7 @@ When completing the "calendar format" step above, consider using one
 of two possible methods: spiceypy.etcal or spiceypy.timout.
 
 Solution
---------
+^^^^^^^^
 
 Solution Meta-Kernel
 
@@ -562,7 +575,7 @@ Obtaining Target States and Positions (getsta)
 .. _task-statement-rs-1:
 
 Task Statement
---------------
+^^^^^^^^^^^^^^
 
 Write a program that prompts the user for an input UTC time string,
 computes the following quantities at that epoch:
@@ -592,7 +605,7 @@ and displays the results. Use the program to compute these quantities at
 .. _learning-goals-rs-1:
 
 Learning Goals
---------------
+^^^^^^^^^^^^^^
 
 Understand the anatomy of an spiceypy.spkezr call. Discover the
 difference between spiceypy.spkezr and spiceypy.spkpos. Familiarity with
@@ -602,7 +615,7 @@ SpiceyPy.
[…]

[…]
