+++ docs/insitu_sensing.rst
@@ -306,7 +306,7 @@ output:
       ET        =     140254384.184625
 
 "UTC to ET" Code
--------------------
+^^^^^^^^^^^^^^^^^
 
 Program "convrt.py":
 
