+++ docs/conf.py
@@ -35,7 +35,12 @@ extensions = [
     "sphinx.ext.autodoc",
     "sphinx_autodoc_typehints",
     "sphinx.ext.viewcode",
+    "autodocsumm",
 ]
+
+autodoc_default_options = {
+    "autosummary": True,
+}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
