+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 numpy
 sphinx_autodoc_typehints
+autodocsumm
