+++ docs/environment.yml
@@ -8,4 +8,3 @@ dependencies:
   - numpy
   - pip:
     - sphinx_autodoc_typehints
-    - autodocsumm
