+++ docs/conf.py
@@ -111,9 +111,7 @@ html_theme = "default"
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
- html_theme_options = {
-     'navigation_depth': 4
- }
+html_theme_options = {"navigation_depth": 4}
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []
