### 2020-02-16 fed37b5
doc heading fixes and adding links to required reading, user guides, … (#352)
* doc heading fixes and adding links to required reading, user guides, and permuted index
* other small fixes
