+++ docs/conf.py
@@ -37,6 +37,7 @@ extensions = [
     "sphinx.ext.viewcode",
 ]
 
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
@@ -110,7 +111,7 @@ html_theme = "default"
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-# html_theme_options = {}
+html_theme_options = {"navigation_depth": 4}
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []
