+++ docs/other_stuff.rst
@@ -62,6 +62,11 @@ These tutorials are available from the NAIF ftp server at JPL:
       https://naif.jpl.nasa.gov/naif/tutorials.html
 
 Required Readings
+^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Required Readings <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html>`_ are also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html.
 
 The Required Reading documents are provided with the Toolkit and are
 located under the "cspice/doc" directory in the CSPICE Toolkit
@@ -78,6 +83,11 @@ installation tree.
       windows.req      The SPICE window data type
 
 The Permuted Index
+^^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Permuted Index <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html>`_ is also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html.
 
 Another useful document distributed with the Toolkit is the permuted
 index. This is located under the "cspice/doc" directory in the C
@@ -88,6 +98,7 @@ discover which SpiceyPy functions perform functions of interest, as well
 as the names of the source files that contain these functions.
 
 SpiceyPy API Documentation
+^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 A SpiceyPy function's parameters specification is available using the
 built-in Python help system.
@@ -125,7 +136,7 @@ the NAIF server at JPL:
 
 ::
 
-      ftp://naif.jpl.nasa.gov/pub/naif/toolkit_docs/Lessons/
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/Lessons/
 
 SpiceyPy Modules Used
 ---------------------
@@ -210,9 +221,15 @@ needing information concerning SPICE are:
        --   Tutorials
 
 Required Reading and Users Guides
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+.. tip::
+   The `Required Readings <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html>`_ are also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html.
+   The `User Guides <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/ug/index.html>`_ are also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/ug/index.html
 
 NAIF Required Reading (\*.req) documents introduce the functionality of
-particular SpiceyPy subsystems:
+particular Spice subsystems:
 
 ::
 
@@ -288,6 +305,7 @@ and Users Guides, accessible from the HTML documentation directory:
          ../cspice/doc/html/index.html
 
 Library Source Code Documentation
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 All SPICELIB and CSPICE source files include usage and design
 information incorporated in a comment block known as the “header.”
@@ -348,6 +366,7 @@ The source code for SpiceyPy products is stored in 'src' sub-directory
 of the main SpiceyPy directory:
 
 API Documentation
+^^^^^^^^^^^^^^^^^^^
 
 The SpiceyPy package is documented in "readthedocs" website:
 
@@ -504,14 +523,14 @@ Lesson 1: Kernel Management with the Kernel Subsystem
 -----------------------------------------------------
 
[…]
