+++ docs/insitu_sensing.rst
@@ -46,6 +46,11 @@ These tutorials are available from the NAIF ftp server at JPL:
       https://naif.jpl.nasa.gov/naif/tutorials.html
 
 Required Readings
+^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Required Readings <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html>`_ are also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/req/index.html.
 
 The Required Reading documents are provided with the Toolkit and are
 located under the "cspice/doc" directory in the CSPICE Toolkit
@@ -62,6 +67,11 @@ installation tree.
       time.req         UTC to ET time conversion
 
 The Permuted Index
+^^^^^^^^^^^^^^^^^^^
+
+.. tip::
+   The `Permuted Index <https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html>`_ is also available on the NAIF website at:
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/info/cspice_idx.html.
 
 Another useful document distributed with the Toolkit is the permuted
 index. This is located under the "cspice/doc" directory in the C
@@ -72,6 +82,7 @@ discover which SpiceyPy functions perform functions of interest, as well
 as the names of the source files that contain these functions.
 
 SpiceyPy API Documentation
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 A SpiceyPy function's parameters specification is available using the
 built-in Python help system.
@@ -115,7 +126,7 @@ the NAIF server at JPL:
 
 ::
 
-      ftp://naif.jpl.nasa.gov/pub/naif/toolkit_docs/Lessons/
+      https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/Lessons/
 
 SpiceyPy Modules Used
 ---------------------
@@ -176,14 +187,14 @@ Step-1: "UTC to ET"
 ------------------------------
 
 "UTC to ET" Task Statement
------------------------------
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Write a program that computes and prints the Ephemeris Time (ET)
 corresponding to "2004-06-11T19:32:00" UTC, as the number of
 ephemeris seconds past J2000, .
 
 "UTC to ET" Hints
---------------------
+^^^^^^^^^^^^^^^^^^^^
 
 Find out what SPICE kernel(s) is(are) needed to support this conversion.
 Reference the "time.req" and/or "Time" tutorial.
@@ -213,7 +224,7 @@ place to look for declaration specification and examples) and output
 print statements.
 
 "UTC to ET" Solution Steps
------------------------------
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Only one kernel file is needed to support this conversion – an LSK file
 "naif0008.tls".
@@ -295,7 +306,7 @@ output:
       ET        =     140254384.184625
 
 "UTC to ET" Code
--------------------
+^^^^^^^^^^^^^^^^^
 
 Program "convrt.py":
 
@@ -345,13 +356,13 @@ Step-2: "SCLK to ET"
 ------------------------------
 
 "SCLK to ET" Task Statement
-------------------------------
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Extend the program from Step-1 to compute and print ET for the following
 CASSINI on-board clock epoch "1465674964.105".
 
 "SCLK to ET" Hints
----------------------
+^^^^^^^^^^^^^^^^^^^^^
 
 Find out what additional (to those already loaded in Step-1) SPICE
[…]

[…]
