### 2020-02-16 #2050 @hayesla
I've also completed my review, and I recommend accepting in its current form.

This is a really great package, and while it was quite a learning curve for me as I have never dealt with SPICE before, the SpiceyPy lessons were extremely helpful!

If I was to suggest things for the future (not in relation to this review) is to perhaps add more simple examples like the Cassini Position Example, and have the functionality to download the example as a `.py` file so users can play around with it.

Also some helper functions about which kernels are required would be such a plus (however I imagine this is a lot of work) but it would certainly make it more user friendly for people not familiar with SPICE.

And one final absolutely minor comment, on the instillation [page](https://spiceypy.readthedocs.io/en/master/installation.html#how-to-install-from-source-for-bleeding-edge-updates) it should be `git clone https://github.com/AndrewAnnex/SpiceyPy.git`
