### 2020-02-17 a284b4c
Update installation.rst
making some changes to install docs suggested at https://github.com/openjournals/joss-reviews/issues/2050
