### 2020-02-17 3c182a7
Update installation.rst
