### 2020-02-17 #2050 @AndrewAnnex
@hayesla @jessie-dotson
Thank you both for reviewing SpiceyPy! 🎉  As you may have seen I have already started incorporating your feedback into the docs and as issues to be addressed. The kernel management and suggestion idea from @hayesla is a goal that other SPICE users and the NAIF IIRC have been looking into, but it is a substantial project unto itself.

Thanks again, and for your suggested improvements! -Andrew
