### 2020-02-19 #2050 @mbobra
@AndrewAnnex We're almost there! Do you mind fixing the following:

* Can you please make your affiliations consistent? Some affiliations includes city, country, and postal code, and others simply include the name of the institute.
* The first reference needs a space between 150 and the following parenthesis.
* Can you please make a new Zenodo release after fixing these issues and make sure the title and author list of the Zenodo release matches that of the JOSS paper?
