### 2020-02-19 #2050 @AndrewAnnex
@mbobra so the 2nd issue regarding the first reference and the parentheses appears to be the default style for rendering the volume and number in the bib entry (e.g. "volume(number)"), and can be seen in the other reference entries. But I also double checked the reference and google scholar seems to set the number to 2018, so I have that change made in a local branch.

I will look into making the affiliations consistent, most likely I will remove the additional location information. FYI Author Ben Pearson is not affiliated with an institution (I have their affiliation listed as "General Public" at the moment). If instead I should locate all the additional information for each author please let me know.

For the zenodo release, is it sufficient to update that information by-hand in the zenodo release via the zenodo website, or should I include a .zenodo.json file in the repo? Note that the new zenodo release will result in version 3.0.2 (v3.0.2) of spiceypy.
