### 2020-02-19 1c9ad9b
removed additional info on author affiliations per editor, fixed number for acton citation
