+++ joss/paper.bib
@@ -23,7 +23,7 @@
   doi = {10.1016/j.pss.2017.02.013},
   issn = {00320633},
   journal = {Planetary and Space Science},
-  number = {January 2017},
+  number = {2018},
   pages = {9--12},
   publisher = {Elsevier Ltd},
   title = {{A look towards the future in the handling of space science mission geometry}},
