### 2020-02-19 201a4cb
removed additional info on author affiliations per editor, fixed number for acton citation (#354)
