### 2020-02-19 #2050 @AndrewAnnex
@mbobra I think the first two issues are now fixed, I have started looking into the zenodo release but the question about the json file still stands, but I will check back in tomorrow
