### 2020-02-19 #2050 @AndrewAnnex
@mbobra I just noticed that we may have to add an author to the author list. Currently the author list includes every contributor (large and small) to the repo who responded prior to the submission to JOSS for review, which was all of the contributors. During the pre-review period but after the software report which listed code contributions, two commits were merged into the master branch, see https://github.com/AndrewAnnex/SpiceyPy/commit/fde56cd41935b15e3e3735db03cc9c0311711d73 and https://github.com/AndrewAnnex/SpiceyPy/commit/c2ed0c05aaa65a373bee8a1f430c70027a8adea5.

At the time I must have concluded that this author did not meet the deadline for inclusion as an author but since a new release is needed it would be more consistent to include them with my permissive policy on authorship. What do you advise in this situation?
