### 2020-02-19 930cb89
* bump to 3.0.2 for joss * added zenodo.json file * updated changelog
