+++ setup.py
@@ -115,7 +115,7 @@ readme.close()
 
 setup(
     name="spiceypy",
-    version="3.0.1",
+    version="3.0.2",
     license="MIT",
     author="Andrew Annex",
     author_email="ama6fy@virginia.edu",
