### 2020-02-19 e9dfdae
added johan as an author to joss paper, updates to readme may not be working correctly
