### 2020-02-19 #2050 @AndrewAnnex
@mbobra I've received permission from the new author so I have went ahead and added their information and the above proof includes their information. If everything looks good I'll merge that change into the master branch and cut a new release 🚀
