### 2020-02-19 805cfff
v3.0.2 bump and changes for zenodo/joss  (#355)
* bump to 3.0.2 for joss
* added zenodo.json file
* updated changelog
* added johan as a coauthor to joss paper
