+++ docs/conf.py
@@ -59,9 +59,9 @@ copyright = "2014-2020, Andrew Annex"
 # built documents.
 #
 # The short X.Y version.
-version = "3.0.1"
+version = "3.0.2"
 # The full version, including alpha/beta/rc tags.
-release = "3.0.1"
+release = "3.0.2"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
