### 2020-02-19 e70a959
small fix to codemeta for new version
