### 2020-02-20 #2050 @mbobra
Thanks for all this! Here are my answers:

* It is okay to have no affiliation. We recommend using 'None' instead of general public, so please change it to 'None.'
* All the other affiliations look good 😄
* It is sufficient to update the Zenodo deposit information by hand, so please go ahead and merge this branch into master (after making the affiliation change), cut a new release, deposit in Zenodo, edit any necessary metadata by hand, and then link the Zenodo DOI in your `README.rst file`.
