### 2020-02-20 #2050 @AndrewAnnex
@mbobra okay great I think we are almost ready then, but annoyingly I just noticed that all of my affiliations end with periods. Should I remove them?
