### 2020-02-20 #2050 @mbobra
> @mbobra okay great I think we are almost ready then, but annoyingly I just noticed that all of my affiliations end with periods. Should I remove them?

Hmm...I just looked at the most recent JOSS publications (https://joss.theoj.org/papers/published) and it looks like none of the affiliations end with periods. So, yes. Sorry :( We're almost there!!!
