### 2020-02-20 b9a6c73
removing periods from affiliations
