### 2020-02-20 #2050 @AndrewAnnex
@mbobra turns out the .zenodo.json file was a bad idea and broke the release hence the delay. But in any case I fixed it!

the Zenodo DOI for SpiceyPy 3.0.2 is 10.5281/zenodo.3676146, note that the version is now v3.0.2.

:shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit: :shipit:
