### 2020-02-21 a12d29d
Update paper.bib
updating some references in JOSS article
