+++ joss/paper.bib
@@ -93,7 +93,6 @@
   author={Albin, Thomas},
   school={University of Stuttgart},
   doi={10.18419/opus-10454},
-  url={https://doi.org/10.18419/opus-10454},
   year={2019}
 }
 
@@ -121,15 +120,14 @@
 }
 
 @inproceedings{wilson:2016,
-  title={Scheduling observations of celestial objects for Earth observing sensor calibration},
-  author={Wilson, Truman and Xiong, Xiaoxiong},
-  booktitle={Sensors, Systems, and Next-Generation Satellites XX},
-  volume={10000},
-  pages={1000011},
-  year={2016},
-  doi={10.1117/12.2240648},
-  url={https://doi.org/10.1117/12.2240648},
-  organization={International Society for Optics and Photonics}
+  doi = {10.1117/12.2240648},
+  year = {2016},
+  month = oct,
+  publisher = {{SPIE}},
+  author = {Truman Wilson and Xiaoxiong Xiong},
+  editor = {Roland Meynart and Steven P. Neeck and Toshiyoshi Kimura and Haruhisa Shimoda},
+  title = {Scheduling observations of celestial objects for {Earth} observing sensor calibration},
+  booktitle = {{Sensors, Systems, and Next-Generation Satellites XX}}
 }
 
 @inproceedings{costa:2018,
