### 2020-02-21 #2050 @kyleniemeyer
Hi @AndrewAnnex, I submitted two small PRs that did some copy-editing:
https://github.com/AndrewAnnex/SpiceyPy/pull/356 and https://github.com/AndrewAnnex/SpiceyPy/pull/357
