### 2020-02-21 e08c56a
Update paper.md (#356)
Fix citation to multiple papers
